Buy/Sell/Match logic updated

+++ spring-brokerage-project/stock-orders/src/main/java/org/brokage/stockorders/dto/AssetDTO.java
@@ -1,11 +1,13 @@
 package org.brokage.stockorders.dto;
 
 
+import java.math.BigDecimal;
+
 public record AssetDTO(
         Long id,
         Long customerId,
 
         String assetName,
-        Long size,
-        Long usableSize
+        BigDecimal size,
+        BigDecimal usableSize
 ) {}
